# Implement the `AESBaseCipher` class and let the `AES128Cipher` and `AES256Cipher` classes extend it