More information is provided in the commit messages. Unit tests are already available and still pass. I highly recommend to review this per commit instead of looking at the full diff and reviewing the second commit with the `?w=1` flag.

The diff of the two classes after the first commit can be found at https://www.diffchecker.com/fBgVIwRm. This clearly shows the amount of code duplication and that basically everything can live in the base class except for `_expandKey`. The `decryptBlock` differences are resolved by taking the one for AES-256 in the base class and setting the `iv` parameter to `null` by default so that for AES-128 nothing changes.

Fixes #5155.