@timvandermeij 2018-02-03
Thank you for the review! I have made all changes and I did a benchmark of the code.

I created the following `benchmark.json` file with two files from the test suite that use the AES-128 cipher (found using `grep -r "AESV2" test/*` and confirmed by placing print statements in the class and opening the files in the viewer):

```
[
  {
    "id": "i9",
    "file": "pdfs/i9.pdf",
    "md5": "ba7cd54fdff083bb389295bc0415f6c5",
    "rounds": 20,
    "type": "load"
  },
  {
    "id": "ichiji.pdf",
    "file": "pdfs/ichiji.pdf",
    "md5": "66b645802d33513cd598886e017392b8",
    "rounds": 20,
    "type": "load"
  }
]
```

Then I ran all benchmarking steps from https://github.com/mozilla/pdf.js/wiki/Benchmarking-your-changes, which I updated to be more step-by-step and to include the `gulp generic` step, which was a really good point from you. Doing so I obtained the following table:

```
-- Grouped By browser, stat --
browser | stat         | Count | Baseline(ms) | Current(ms) | +/- |    %  | Result(P<.05)
------- | ------------ | ----- | ------------ | ----------- | --- | ----- | -------------
Firefox | Overall      |   120 |           55 |          54 |  -1 | -1.69 |              
Firefox | Page Request |   120 |            3 |           3 |   0 | -4.26 |              
Firefox | Rendering    |   120 |           52 |          52 |  -1 | -1.59 |
```

Therefore, we can conclude that this change has no negative performance impact.